requirements.txt
Flask
gtts
pydub
requests
python-dotenv
flask_cors